分布式session登录 + 异常全局处理

+++ src/main/java/top/tywang/seckill/dao/UserDao.java
@@ -9,7 +9,7 @@ import top.tywang.seckill.domain.User;
 @Mapper
 public interface UserDao {
     @Select("select * from user where id = #{id}")
-    public User getById(@Param("id") int id);
+    public User getById(@Param("id") long id);
 
     @Insert("insert into user(id, name)values(#{id}, #{name})")
     public int insert(User user);
